Strange audio starting timing workaround
Better than a hard-coded time to wait for the audio to begin. The technique checks the memory the moment the ffplay EXE is initialised and waits for an increase in memory usage, indicating the audio has been loaded. Will use this for now and hopefully it works well even in difficult situations e.g. lag.