Resize table columns, specialised table column class.
More appropriate sizing of columns. Name needs to be the widest of course. Still, supporting 100 characters in one cell is overkill. Will need to add an ellipsis between the start and end of very long playlist names, just like long audio paths having an ellipsis between the start and end of the path.