Very minor refinements
Add Playlist section in view playlist window. Add a menu for the import folder toplevel. Fix fatal bug which prevented creating a playlist. Now, if the playlist form is invoked from the view playlists window, go back to the view playlists window instead of the home screen, as it makes more sense.